+++ CHANGELOG.rst
@@ -15,7 +15,9 @@ Added
 - DMRG parameter ``'orthogonal_to'`` allows to calculate excited states for finite systems.
 - possibility to change the number of charges after creating LegCharges/Arrays
 - more general way to specify the order of sites in a :class:`tenpy.models.lattice.Lattice`.
-- new :class:`tenpy.models.lattice.Honeycomb`
+- new :class:`tenpy.models.lattice.Honeycomb` lattice
+- a way to specify nearest neighbor couplings in a :class:`~tenpy.models.lattice.Lattice`, 
+  along with methods to count the number of nearest neighbors for sites in the bulk.
 
 Changed
 ^^^^^^^